+++ bullet.h
@@ -9,13 +9,13 @@ class bullet : public QObject, public QGraphicsPixmapItem
 {
     Q_OBJECT
 public:
-    bullet(QGraphicsScene *mainwin, const int i, double X, double Y, int count);
+    bullet(QGraphicsScene *mainwin, const int i, float X, float Y, int count);
     QGraphicsPixmapItem *blt;
     void con();
     mainbullet main_bul;
 
 protected:
-    double X_axis, Y_axis;
+    float X_axis, Y_axis;
     int type;
     QPixmap *bullet_pix_type_1;
     QPixmap *bullet_pix_type_2;
